Discover_eHFO/requirements.txt
torchvision==0.7.0
scipy==1.5.0
matplotlib==3.3.2
pandas==0.25.1
tqdm==4.47.0
scikit_image==0.17.2
numpy==1.18.5
torch==1.6.0
seaborn==0.10.1
pyEDFlib==0.1.18
matlab==0.1
Pillow==8.2.0
scikit_learn==0.24.1
skimage==0.0
